More detailed compiler options, warning cleanup
Fix is_move_valid bug
Inline bitboard ops

+++ src/hashing/poskey.cpp
@@ -56,7 +56,7 @@ uint64_t PosKey::rand_64() {
     std::mt19937_64 e2(rd());
 
     std::uniform_int_distribution<long long int> dist(std::llround(std::pow(2, 61)), std::llround(std::pow(2, 62)));
-    return dist(e2);
+    return static_cast<uint64_t>(dist(e2));
 }
 
 void PosKey::init_hash_keys() {
