+++ src/hashing/pvtable.h
@@ -23,7 +23,7 @@ public:
         PVEntry() : age(0), flag(EntryFlag::HFNONE) {};
         PVEntry(const PosKey& key,
                 const SearchMove &m,
-                int p_depth,
+                uint8_t p_depth,
                 EntryFlag p_flag) {
             posKey = key;
             move = m;
@@ -35,14 +35,14 @@ public:
 
         PosKey posKey;
         SearchMove move;
-        int depth = 0;
-        int age;
+        uint8_t depth = 0;
+        uint8_t age;
         EntryFlag flag = EntryFlag::HFNONE;
         bool full = false;
     };
 
     // Default size of 64 MB.
-    explicit PVTable(int size = 0x100000 * 64);
+    explicit PVTable(size_t size = 0x100000 * 64);
 
     PVTable(const PVTable& rhs);
     PVTable& operator=(const PVTable& rhs);
@@ -52,21 +52,21 @@ public:
     std::optional<PVEntry> probe(const PosKey& key);
     std::optional<SearchMove> probe_move(const PosKey& key, int alpha, int beta, int depth);
 
-    PVTable resize(int new_size);
+    PVTable resize(size_t new_size);
     void prep_search();
 
 private:
-    [[nodiscard]] int hash(const PosKey& key) const;
+    [[nodiscard]] uint64_t hash(const PosKey& key) const;
 
     std::unique_ptr<PVEntry[]> table;
-    int entries;
-    int entries_full = 0;
+    uint64_t entries;
+    uint64_t entries_full = 0;
 
-    int new_write = 0;
-    int overwrite = 0;
-    int hit = 0;
-    int cut = 0;
-    int current_age = 0;
+    uint64_t new_write = 0;
+    uint64_t overwrite = 0;
+    uint64_t hit = 0;
+    uint64_t cut = 0;
+    uint8_t current_age = 0;
 };
